Service/requirements.txt
djangorestframework
django-contrib-comments
django
